refactor: 폴더명 snake_case 로 수정
