+++ lib/presentation/pages/guide_package_detail_page/widgets/detail_schedule_list_view_model.dart
