+++ lib/presentation/pages/guide_package_detail_page/widgets/package_detail_header.dart
