+++ lib/presentation/pages/guide_package_detail_page/widgets/schedule_detail_item.dart
